requirements.txt
dash
pandas
plotly
dash-mantine-components
databricks-sdk